Fixed: Removed value bindings for Capp Lint in XCodeCapp .xib file
Previously the "availability" bindings for the Code Style notifications and binding window in the preference pane were set to non-available variables. This resulted in the permanently deactivated checkboxes for these preferences.

This commit removes the availability bindings, restoring the checkbox functionality in the preference pane.

Additionally, this commit changes the "XCCAutoShowNotificationOnCappLint" default variable to active by default to match all other preferences.