Fixed: Crash of press tool when having a CPWebView in a CIB
This fixes the crash of the press tool when a CPWebView is included into a CIB.

In this case, the view is not added at all to the DOM so _iframe is not created but the scroll change method is anyway called. This fix makes sure that the _iframe is not accessed in this case.