New: A private notification is sent to every subviews when the bounds or the frame of a view changed
Previously, when the frame or the bounds of a view changed, the subviews of this view were not notified by this event.
Now they are, every subview will send a private notification that one of their parent's frame/bounds has changed.

This new feature fixes the automatic positioning of a popover when the targeted view's frame change.