New: compilation gives more informations when getting an error or warning
Previously, when the compilation of a project failed, we did not have so many informations about the cause of the problem.
Now, when gcc returns an error or a warning, we print in the console. This help to know what is the issue of the build we would like to have.