Depreciation warning for -reloadContent
Also removed the reloadContent test.

Previously, the collection view was keeping a reference of the content
provided.  If the later changed externally, we had to notify the
collection view that the layout needed to be updated using
-reloadContent.
After this PR, the receiver makes a copy of the content (like cocoa).
The only way to set the content (and update the layout) is now to call
setContent, making reloadContent useless. Internally, the old and new
content are compared and the layout is updated only when
insertions/deletions/moves are needed.

This change is more in line with cocoa and is also needed for items
animations.